requirements.txt
streamlit>=1.32.0
anthropic>=0.86.0
moviepy==1.0.3
Pillow>=10.0.0
PyYAML>=6.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
gTTS>=2.3.2
tqdm>=4.65.0
